refactor: move AddUserPreference to user-preference folder
