+++ src/domain/contracts/user-preference/index.ts
@@ -1,0 +1,1 @@
+export * from './add-user-preference'
