+++ src/main/factories/infra/db/prisma/user-preference/add-user-preference-prisma-repo-factory.ts
@@ -1,5 +1,5 @@
 import { AddUserPreferencePrismaRepo } from '@/infra/db/prisma/data/user-preference/add-user-preference/add-user-preference-prisma-repo'
-import { type AddUserPreferenceRepo } from '@/usecases/contracts/db/user'
+import { type AddUserPreferenceRepo } from '@/usecases/contracts/db/user-preference'
 
 export const makeAddUserPreferencePrismaRepo = (): AddUserPreferenceRepo => {
   return new AddUserPreferencePrismaRepo()
