+++ src/main/factories/usecases/user-preference/add-user-preference-usecase-factory.ts
@@ -1,7 +1,7 @@
-import { type AddUserPreference } from '@/domain/contracts/user/add-user-preference'
 import { makeFindUserPreferenceByIdPrismaRepo } from '../../infra/db/prisma/user-preference/find-preference-by-id-prisma-repo-factory'
 import { AddUserPreferenceUsecase } from '@/usecases/implementations/user-preference/add-user-preference/add-user-preference-usecase'
-import { makeAddUserPreferencePrismaRepo } from '../../infra/db/prisma/user/add-user-preference-prisma-repo-factory'
+import { makeAddUserPreferencePrismaRepo } from '../../infra/db/prisma/user-preference/add-user-preference-prisma-repo-factory'
+import { type AddUserPreference } from '@/domain/contracts/user-preference'
 
 export const makeAddUserPreferenceUseCase = (): AddUserPreference => {
   return new AddUserPreferenceUsecase(
