+++ src/main/routes/user/user-route.ts
@@ -2,7 +2,7 @@ import { adaptRoute } from '@/main/factories/adapters'
 import { makeSaveUserPreferenceDayPeriodController } from '@/main/factories/controllers/user-preference-day-period/save-user-preference-day-period-controller-factory'
 import { makeSaveUserPreferenceGamePlaceController } from '@/main/factories/controllers/user-preference-game-place/save-user-preference-game-place-controller-factory'
 import { makeSaveUserSocialMediaController } from '@/main/factories/controllers/user-social-media/save-user-social-media-controller-factory'
-import { makeAddUserPreferenceController } from '@/main/factories/controllers/user/add-user-preference-controller-factory'
+import { makeAddUserPreferenceController } from '@/main/factories/controllers/user-preference/add-user-preference-controller-factory'
 import { makeUpdateUserController } from '@/main/factories/controllers/user/update-user-controller-factory'
 import { makeUpdateUserPreferenceController } from '@/main/factories/controllers/user/update-user-preference-controller-factory'
 import { Controller, Patch, Post, Put, Req, Res } from '@nestjs/common'
