+++ src/usecases/contracts/db/user-preference/add-user-preference-repo.ts
@@ -1,0 +1,5 @@
+import { type UserPreferenceModel } from '@/domain/models'
+
+export interface AddUserPreferenceRepo {
+  execute: (data: UserPreferenceModel) => Promise<void>
+}
