+++ src/usecases/contracts/db/user-preference/index.ts
@@ -1,1 +1,2 @@
 export * from './find-user-preference-by-id-repo'
+export * from './add-user-preference-repo'
