--- src/usecases/contracts/db/user/add-user-preference-repo.ts
@@ -1,5 +1,0 @@
-import { type AddUserPreferenceData } from '@/domain/contracts/user/add-user-preference'
-
-export interface AddUserPreferenceRepo {
-  execute: (data: AddUserPreferenceData) => Promise<void>
-}
