+++ src/usecases/implementations/user-preference/add-user-preference/add-user-preference-usecase.spec.ts
@@ -1,9 +1,14 @@
-import { type AddUserPreferenceData } from '@/domain/contracts/user/add-user-preference'
+import { type AddUserPreferenceData } from '@/domain/contracts/user-preference'
 import { ExistentUserPreferenceError } from '@/domain/errors'
 import { type UserPreferenceModel } from '@/domain/models'
-import { type AddUserPreferenceRepo } from '@/usecases/contracts/db/user'
-import { type FindUserPreferenceByIdRepo } from '@/usecases/contracts/db/user-preference'
+import type { AddUserPreferenceRepo, FindUserPreferenceByIdRepo } from '@/usecases/contracts/db/user-preference'
 import { AddUserPreferenceUsecase } from './add-user-preference-usecase'
+
+const makeFakeAddUserPreferenceData = (): AddUserPreferenceData => ({
+  userId: 'any_user_id',
+  frequency: 'daily',
+  activeType: 'gameMaster'
+})
 
 const makeFakeUserPreferenceModel = (): UserPreferenceModel => ({
   id: 'any_user_id',
@@ -22,7 +27,7 @@ const makeFindUserPreferenceByIdRepo = (): FindUserPreferenceByIdRepo => {
 
 const makeAddUserPreferenceRepo = (): AddUserPreferenceRepo => {
   class AddUserPreferenceRepoStub implements AddUserPreferenceRepo {
-    async execute (data: AddUserPreferenceData): Promise<void> {
+    async execute (data: UserPreferenceModel): Promise<void> {
       await Promise.resolve()
     }
   }
@@ -46,23 +51,23 @@ describe('AddUserPreferenceUsecase', () => {
   it('Should call FindUserPreferenceByIdRepo with correct values', async () => {
     const { sut, findUserPreferenceByIdRepoStub } = makeSut()
     const executeSpy = jest.spyOn(findUserPreferenceByIdRepoStub, 'execute')
-    await sut.perform(makeFakeUserPreferenceModel())
+    await sut.perform(makeFakeAddUserPreferenceData())
     expect(executeSpy).toHaveBeenCalledWith('any_user_id')
   })
 
-  it('Should return ExistentUserPreferencesError if FindUserPreferenceByIdRepo returns a preference', async () => {
+  it('Should return ExistentUserPreferencesError if FindUserPreferenceByIdRepo returns a UserPreference', async () => {
     const { sut, findUserPreferenceByIdRepoStub } = makeSut()
     jest.spyOn(findUserPreferenceByIdRepoStub, 'execute').mockReturnValueOnce(
       Promise.resolve(makeFakeUserPreferenceModel())
     )
-    const result = await sut.perform(makeFakeUserPreferenceModel())
+    const result = await sut.perform(makeFakeAddUserPreferenceData())
     expect(result.value).toEqual(new ExistentUserPreferenceError('any_user_id'))
   })
 
   it('Should call AddUserPreferenceRepo with correct values', async () => {
     const { sut, addUserPreferenceRepoStub } = makeSut()
     const executeSpy = jest.spyOn(addUserPreferenceRepoStub, 'execute')
-    await sut.perform(makeFakeUserPreferenceModel())
+    await sut.perform(makeFakeAddUserPreferenceData())
     expect(executeSpy).toHaveBeenCalledWith(makeFakeUserPreferenceModel())
   })
 
@@ -71,13 +76,13 @@ describe('AddUserPreferenceUsecase', () => {
     jest.spyOn(addUserPreferenceRepoStub, 'execute').mockReturnValueOnce(
       Promise.reject(new Error())
     )
-    const promise = sut.perform(makeFakeUserPreferenceModel())
+    const promise = sut.perform(makeFakeAddUserPreferenceData())
     await expect(promise).rejects.toThrow()
   })
 
   it('Should return right result on success', async () => {
     const { sut } = makeSut()
-    const result = await sut.perform(makeFakeUserPreferenceModel())
+    const result = await sut.perform(makeFakeAddUserPreferenceData())
     expect(result.isRight()).toBe(true)
   })
 })
