+++ src/usecases/implementations/user-preference/add-user-preference/add-user-preference-usecase.ts
@@ -1,8 +1,7 @@
-import type { AddUserPreference, AddUserPreferenceData, AddUserPreferenceResponse } from '@/domain/contracts/user/add-user-preference'
+import type { AddUserPreference, AddUserPreferenceData, AddUserPreferenceResponse } from '@/domain/contracts/user-preference'
 import { ExistentUserPreferenceError } from '@/domain/errors'
 import { left, right } from '@/shared/either'
-import type { AddUserPreferenceRepo } from '@/usecases/contracts/db/user'
-import type { FindUserPreferenceByIdRepo } from '@/usecases/contracts/db/user-preference'
+import type { AddUserPreferenceRepo, FindUserPreferenceByIdRepo } from '@/usecases/contracts/db/user-preference'
 
 export class AddUserPreferenceUsecase implements AddUserPreference {
   constructor (
@@ -11,12 +10,12 @@ export class AddUserPreferenceUsecase implements AddUserPreference {
   ) {}
 
   async perform (data: AddUserPreferenceData): Promise<AddUserPreferenceResponse> {
-    const { id } = data
+    const { userId: id, ...otherData } = data
     const userPreferenceExists = await this.findUserPreferenceByIdRepo.execute(id)
     if (userPreferenceExists) {
       return left(new ExistentUserPreferenceError(id))
     }
-    await this.addUserPreferenceRepo.execute(data)
+    await this.addUserPreferenceRepo.execute({ id, ...otherData })
     return right(null)
   }
 }
